+++ docs/examples/spam/src/__init__.py
@@ -1,0 +1,1 @@
+from ._spam import add  # noqa: F401